+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.52.0-alpha.30](https://github.com/SocialGouv/jardinmental/compare/v1.52.0-alpha.29...v1.52.0-alpha.30) (2025-08-27)
+
+
+### Bug Fixes
+
+* **indicator:** replace usage of name as main id among all files ([#364](https://github.com/SocialGouv/jardinmental/issues/364)) ([d451fbc](https://github.com/SocialGouv/jardinmental/commit/d451fbcc84a38620305c0d2e5ea8523e37277a31))
+
+# [1.52.0-alpha.29](https://github.com/SocialGouv/jardinmental/compare/v1.52.0-alpha.28...v1.52.0-alpha.29) (2025-08-27)
+
+
+### Bug Fixes
+
+* **indicator:** display the right semantic label, display the right colors ([#365](https://github.com/SocialGouv/jardinmental/issues/365)) ([566a2b6](https://github.com/SocialGouv/jardinmental/commit/566a2b6497fddc4e82fdf2a45134927575015987))
+
 # [1.52.0-alpha.28](https://github.com/SocialGouv/jardinmental/compare/v1.52.0-alpha.27...v1.52.0-alpha.28) (2025-08-26)